Comment out methods that attach SCORM object to the window
We have had issues with SCORM packages that get confused when both window objects are attached (API and API_1484_11).

The current solution is to comment out that code so that we can implement it manually on our side to only active the correct SCORM API version for the content we are embedding.